Move wallbox base entity to its own file
