+++ homeassistant/components/wallbox/lock.py
@@ -9,13 +9,14 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import PlatformNotReady
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import InvalidAuth, WallboxCoordinator, WallboxEntity
+from . import InvalidAuth, WallboxCoordinator
 from .const import (
     CHARGER_DATA_KEY,
     CHARGER_LOCKED_UNLOCKED_KEY,
     CHARGER_SERIAL_NUMBER_KEY,
     DOMAIN,
 )
+from .entity import WallboxEntity
 
 LOCK_TYPES: dict[str, LockEntityDescription] = {
     CHARGER_LOCKED_UNLOCKED_KEY: LockEntityDescription(
